// frontend/src/app/core/generated/circabc/model/informationPage.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InformationPage { 
    url?: string;
    adapt?: boolean;
    displayOldInformation?: boolean;
    /**
     * Representation of the permissions given to the user making the call. To know what are the permissions for the specific node, the permission definition may vary depending on its type or service. 
     */
    permissions?: { [key: string]: string; };
}